requirements.txt
streamlit>=1.32.0
ultralytics>=8.3.0
opencv-python-headless>=4.8.0.74
pillow>=10.2.0
numpy>=1.26.4
torch>=2.2.2
torchvision>=0.17.2
pyyaml>=6.0.1
